refactor : ErrorMessage를 삭제하고 StringConst를 util 패키지에서 View 패키지로 이동
