+++ src/main/java/racingcar/controller/GameController.java
@@ -3,7 +3,7 @@ package racingcar.controller;
 import racingcar.model.CarList;
 import racingcar.model.RandomNumberList;
 import racingcar.util.IntegerConst;
-import racingcar.util.StringConst;
+import racingcar.view.StringConst;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -26,7 +26,7 @@ public class GameController {
         for (int nowTurn = IntegerConst.ZERO.getValue(); nowTurn < totalTurn; nowTurn++) {
             moveCarList(carList);
             OutputView.displayCarPosition(carList.getNameList(), carList.getEachCarPosition(),
-                StringConst.POSITION_DELEIMIER.getValue());
+                StringConst.POSITION_DELIMITER.getValue());
         }
         OutputView.displayWinner(carList.getWinnerNames());
     }
